+++ tests/testthat/test-tables.R
@@ -25,6 +25,15 @@ test_that("stages_table() works as expected", {
 })
 
 
+test_that("stages_table() includes eta in engr units", {
+  t1 <- load_eeu_data() %>% 
+    rebound_analysis() %>% 
+    stages_table()
+  expect_true(!is.null(t1))
+  
+})
+
+
 test_that("rebound_results_table() works as expected", {
   t1 <- load_eeu_data() %>% 
     rebound_analysis() %>%